Feat: 카테고리별 조회 페이지 SSR & SWR 적용
